rename solution name
