+++ InterviewTestProject/Startup.cs
@@ -1,4 +1,4 @@
-using InverviewProject.Services;
+using InterviewTestProject.Services;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
@@ -11,7 +11,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace InverviewProject
+namespace InterviewTestProject
 {
     public class Startup
     {
